+++ skore/tests/unit/displays/roc_curve/test_cross_validation.py
@@ -14,10 +14,15 @@ def test_legend_binary_classification(
     legend = ax[0].get_legend()
     assert legend is not None
     legend_texts = [text.get_text() for text in legend.get_texts()]
-    plot_data = display.frame(with_roc_auc=True)
-    roc_auc = plot_data["roc_auc"]
-    assert legend_texts[0] == f"AUC={roc_auc.mean():.2f}±{roc_auc.std():.2f}"
-    assert len(legend_texts) == 2
+    labels = display.roc_curve["label"].cat.categories
+    for label_idx, label in enumerate(labels):
+        plot_data = display.frame(with_roc_auc=True)
+        roc_auc = plot_data.query(f"label == {label}")["roc_auc"]
+        assert (
+            legend_texts[label_idx] == f"{label} (AUC={roc_auc.mean():.2f}"
+            f"±{roc_auc.std():.2f})"
+        )
+    assert len(legend_texts) == len(labels) + 1
     assert legend_texts[-1] == "Chance level (AUC = 0.5)"
 
 
@@ -48,7 +53,7 @@ def test_legend_multiclass_classification(
 @pytest.mark.parametrize(
     "fixture_name, valid_values",
     [
-        ("cross_validation_reports_binary_classification", ["None", "auto"]),
+        ("cross_validation_reports_binary_classification", ["None", "auto", "label"]),
         (
             "cross_validation_reports_multiclass_classification",
             ["None", "auto", "label"],
@@ -72,7 +77,7 @@ def test_invalid_subplot_by(fixture_name, valid_values, request):
 @pytest.mark.parametrize(
     "fixture_name, subplot_by_tuples",
     [
-        ("cross_validation_reports_binary_classification", [(None, 0)]),
+        ("cross_validation_reports_binary_classification", [(None, 0), ("label", 2)]),
         (
             "cross_validation_reports_multiclass_classification",
             [("label", 3)],